Create .gitignore and Dockerfile
A new .gitignore file is created in the .idea directory to handle ignored files. This should help avoid committing unnecessary or temporary files to the repository. A Dockerfile is also added, using the latest version of Ubuntu, with 'marlisumarli' as the author. The ENTRYPOINT is defined to run a command when a Docker container starts.